refactor: Refactor of gui structure, including: FrequencyComponentWidget, GlobablStyles etc.

+++ src/core/AudioAnalyser/AudioAnalyser.hpp
@@ -17,7 +17,7 @@ public:
     {
         float amplitude;
         float phase;
-        float frequency;
+        int frequency;
     };
 
     AudioAnalyser();
